Split up driver.py
